Rename the compat package to contrib
It will be the package with all optional dependencies
