+++ setup.py
@@ -12,7 +12,7 @@ setuptools.setup(
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/petee-d/stereotype",
-    packages=['stereotype', 'stereotype.fields', 'stereotype.compat', 'examples'],
+    packages=['stereotype', 'stereotype.fields', 'stereotype.contrib', 'examples'],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
